refactor: decoupling item structures to pull into actor
